Moves InitializeNeoPlugin to plugin-state to remove dependency cycle
