+++ src/utils/codap-utils.ts
@@ -1,98 +1,7 @@
-import { addDataContextChangeListener, codapInterface, getCaseByID, initializePlugin, sendMessage }
+import { codapInterface, sendMessage }
   from "@concord-consortium/codap-plugin-api";
-// import {
-//   kPinDataContextName, kPinLatAttributeName, kPinLongAttributeName, kPinColorAttributeName,
-//   kPluginName, kInitialDimensions, kVersion,
-//   kDataContextName, kMapPinsCollectionName, kOneMonthInSeconds,
-//   kMapName, kSliderComponentName, kChartGraphName, kXYGraphName,
-// } from "../data/constants";
-import { kPluginName, kInitialDimensions, kVersion, kOneMonthInSeconds, kMapName,
-  kMapPinsCollectionName, kPinColorAttributeName, kPinDataContextName, kPinLatAttributeName, kPinLongAttributeName,
-  kDataContextName, kSliderComponentName, kChartGraphName, kXYGraphName } from "../data/constants";
-import { IMapPin, pluginState } from "../models/plugin-state";
-
-
-export async function initializeNeoPlugin() {
-  initializePlugin({ pluginName: kPluginName, version: kVersion, dimensions: kInitialDimensions });
-
-  // Create the pin dataset
-  await sendMessage("create", `dataContext`, {
-    name: kPinDataContextName,
-    collections: [
-      {
-        name: "Map Pins",
-        attrs: [
-          { name: kPinLatAttributeName, type: "numeric" },
-          { name: kPinLongAttributeName, type: "numeric" },
-          { name: kPinColorAttributeName, type: "color" }
-        ]
-      }
-    ]
-  });
-
-  // Create map if it doesn't exist
-  await createOrUpdateMap("Map");
-
-  // See if there are any existing pins
-  pluginState.updatePins();
-
-  // Set up a listener for changes to the pin dataset
-  addDataContextChangeListener(kPinDataContextName, notification => {
-    const { operation } = notification.values;
-
-    if (["createCases", "deleteCases", "updateCases"].includes(operation)) {
-      pluginState.updatePins();
-    }
-  });
-  // Set up a listener for pin selection
-  addDataContextChangeListener(kPinDataContextName, async notification => {
-    const { operation, result } = notification.values;
-    if (operation === "selectCases" && result.success) {
-      const selectedPins = await getSelectionList(kPinDataContextName);
-      const selectedPinValues: IMapPin[] = await Promise.all(
-        selectedPins.map(async (pin: any) => {
-          const pinItem = await getCaseByID(kPinDataContextName, pin.caseID);
-          if (pinItem.success) {
-            const pinValues = pinItem.values;
-            const pinCase = (pinValues as any).case;
-            return {
-              id: pinCase.id,
-              lat: pinCase.values.pinLat,
-              long: pinCase.values.pinLong,
-              color: pinCase.values.pinColor,
-            };
-          }
-          return null;
-        })
-      );
-      pluginState.setSelectedPins(selectedPinValues);
-    }
-  });
-
-  // Set up a listener for case selection
-  addDataContextChangeListener(kDataContextName, async notification => {
-    const { operation, result } = notification.values;
-    if (operation === "selectCases" && result.success) {
-      const selectedCases = await getSelectionList(kDataContextName);
-      const selectedPinCases = selectedCases.filter((sCase: any) => sCase.collectionName === kMapPinsCollectionName);
-      const selectedCaseValues: any[] = await Promise.all(
-        selectedPinCases.map(async (sCase: any) => {
-          const caseItem = await getCaseByID(kDataContextName, sCase.caseID);
-          if (caseItem.success) {
-            const caseValues = caseItem.values;
-            return {
-              id: caseValues.id,
-              label: caseValues.case.values.label,
-              pinColor: caseValues.case.values.pinColor,
-            };
-          }
-          return null;
-        })
-      );
-      pluginState.setSelectedCases(selectedCaseValues);
-    }
-  });
-}
+import { kOneMonthInSeconds, kMapName, kMapPinsCollectionName, kDataContextName, kSliderComponentName,
+  kChartGraphName, kXYGraphName } from "../data/constants";
 
 export async function createOrUpdateMap(title: string, url?: string): Promise<void> {
   const mapProps: Record<string, any> = {
